+++ src/main/java/com/jaquadro/minecraft/storagedrawers/api/framing/IFramedBlockEntity.java
@@ -1,7 +1,5 @@
 package com.jaquadro.minecraft.storagedrawers.api.framing;
 
-import com.jaquadro.minecraft.storagedrawers.block.tile.tiledata.MaterialData;
-
 public interface IFramedBlockEntity
 {
     IFramedMaterials material ();